refactor(ocap-kernel): reorganize code by domain into vats, liveslots, and garbage-collection directories
