+++ packages/ocap-kernel/src/garbage-collection/garbage-collection.ts
