+++ packages/ocap-kernel/src/liveslots/kernel-marshal.test.ts
