+++ packages/ocap-kernel/src/liveslots/kernel-marshal.ts
