+++ packages/ocap-kernel/src/liveslots/meter-control.ts
